target/riscv: Allow setting ISA extensions via CPU props
This patch allows us to enable/disable the RISC-V ISA extensions from
the QEMU command line. This works with the rv32 and rv64 machines. The
idea is that in the future we can now add extensions and leave them
disabled by default until enabled by the user.

---CHANGELOG---

Based-on: <cover.1555726824.git.alistair.francis@wdc.com>

This is based on my previous series: "RISC-V: Add properties to the
CPUs" which has been merged into the RISC-V tree.

---END---

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>