COVER: RISC-V: Hypervisor prep work part 2
---CHANGELOG---
PUBLIC

The first three patches are ones that I have pulled out of my original
Hypervisor series at an attempt to reduce the number of patches in the
series.

These three patches all make sense without the Hypervisor series so can
be merged seperatley and will reduce the review burden of the next
version of the patches.

The fource patch is a prep patch for the new v0.4 Hypervisor spec.

The final patch is unreleated to Hypervisor that I'm just slipping in
here because it seems easier then sending it by itself.

v2:
 - Small corrections based on feedback
 - Remove the CSR permission check patch

---END---

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>